+++ src/SixLabors.Fonts/Native/CoreText.cs
@@ -4,16 +4,17 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace SixLabors.Fonts.Native;
-
-internal static class CoreText
+namespace SixLabors.Fonts.Native
 {
-    private const string CoreTextFramework = "/System/Library/Frameworks/CoreText.framework/Versions/A/CoreText";
+    internal static class CoreText
+    {
+        private const string CoreTextFramework = "/System/Library/Frameworks/CoreText.framework/Versions/A/CoreText";
 
-    /// <summary>
-    /// Returns an array of font URLs.
-    /// </summary>
-    /// <returns>This function returns a retained reference to a <c>CFArray</c> of <c>CFURLRef</c> objects representing the URLs of the available fonts, or <c>NULL</c> on error. The caller is responsible for releasing the array.</returns>
-    [DllImport(CoreTextFramework, CharSet = CharSet.Ansi, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
-    public static extern IntPtr CTFontManagerCopyAvailableFontURLs();
+        /// <summary>
+        /// Returns an array of font URLs.
+        /// </summary>
+        /// <returns>This function returns a retained reference to a <c>CFArray</c> of <c>CFURLRef</c> objects representing the URLs of the available fonts, or <c>NULL</c> on error. The caller is responsible for releasing the array.</returns>
+        [DllImport(CoreTextFramework, CharSet = CharSet.Ansi, CallingConvention = CallingConvention.Cdecl, ExactSpelling = true)]
+        public static extern IntPtr CTFontManagerCopyAvailableFontURLs();
+    }
 }
